requirements.txt
flask
python-dotenv
openai
tiktoken
sentence-transformers
faiss-cpu
markdown
